Prefer `SetValue()` over `IsDirty`
In anticipation of returning `AttributeValue` back to an immutable, read-only value object, we're preferring the use of `SetValue()` over `IsDirty`. `SetValue()` currently still writes to properties, but it will be easier to update the `SetValue()` code to handle copying and modifying the value object than it will be to decentralize that code each place where we need to modify an `AttributeValue`'s properties.

Not all instances are covered here, as some need to be first addressed by the `SetValue()` code before we remove this dependency. For instance, see the setting of `IsDirty` and `LastModified` in `EnforceBusinessLogic()`.

(This is also a secondary benefit of moving back to an immutable model: it helps encourage use of `SetValue()` over working with `AttributeValue` instances directly.)