Introduced `IsExternalInit` as polyfill
In C# 9.0 and the latest versions of .NET 3.1, there is an `IsExternalInit` class which permits init-only setters. We'll want to take advantage of those in order to move `AttributeValue` back to an immutable value object, as that allows its properties to be initialized without each one being represented by a constructor parameter.

Since .NET Standard 2.0 (e.g., .NET Framework 4.8) doesn't have this class, however, we need to manually establish it as a type of "polyfill". That said, there's no actual functionality needed here; we just need an empty internal class to represent this type.