Ensured initial `AttributeValue` tracked if inserted directly
When business logic is enforced, we don't expect the values on the `AttributeValue` to change—and especially those that aren't understood by the business logic, such as `IsDirty` or `LastModified`. The only goal of enforcing business logic is to validate the `value` and ensure any internal state is maintained appropriately. As such, after business logic is enforced, we expect the original `IsDirty` and `LastModified` values to be maintained. In fact, we generally expect that we'll just persist the _original_ `AttributeValue` object with whatever properties were set on it.

This is what the new `BusinessLogicCache` manages, by maintaining a copy of that original `AttributeValue` instance. If a record exists in the cache, we know that it is currently doing its round trip through the business logic. This worked fine so long as an implementer called `SetValue()`. But it didn't work if the user went around `SetValue()` by e.g. adding the `AttributeValue` directly. In this case, that original value would not be set in the `BusinessLogicCache`.

By accounting for this in `EnforceBusinessLogic()`, we ensure that this scenario is accounted for, in addition to calling `SetValue()` first (either directly, or through a property accessor, such as `Topic.Key`).