Introduced unit test to validate business logic honors parameters
Parameters such as `IsDirty` should be honored even if the normal business logic would have changed their values. So, for instance, if I call `SetValue()` with `isDirty=false`, and then it calls the `Key` property as part of `EnforceBusinessLogic()`, the `IsDirty` property should remain false even if the `Key` property changed the value. By explicitly setting the `isDirty` parameter or `AttributeValue.IsDirty` property, we are sending clear instructions that we expect that value to be set to false for this transaction.

A bug in this behavior was identified and resolved in the previous commit (df720f8), addressing a scenario where this wasn't honored if a new `AttributeValue` object was added directly to the collection. This unit test evaluates both that scenario as well as the `SetValue()` scenario to ensure that both operate as expected.