Handled scenario where business logic changes value
If `SetValue()` is called or a new `AttributeValue` is added to the `AttributeValueCollection`, and the `AttributeValue.Key` maps to a property on the associated `Topic` instance, then that property will be called in order to enforce business logic. During this process, we normally don't expect the state of the `AttributeValue` to change; i.e., we might expect the property to change its internal state, or even to throw an exception if the value is invalid, but we wouldn't expect it to change the _value_.

That's probably a fine assumption. But it isn't guaranteed. And while it's not usually a recommended design choice, this update ensures that if it _does_ happen, then that change is honored. This helps avoid scenarios that, while rare, would be really difficult to isolate otherwise, and might even open up for downstream bugs.

For instance, a property might clean up the `value` to remove unnecessary or even dangerous content (such as user submitted text containing JavaScript or SQL in an effort to implement an injection exploit). Usually, we'd use a method to do that prior to setting the property, or handle it as part of a model binder, but it's just one scenario where this _might_ occur.

Regardless, it doesn't hurt to check, and avoid potential confusion if this scenario does occur. So we're evaluating if the new `value` has been modified after `EnforceBusinessLogic()` and, if so, updating `AttributeValue.Value` to the new value.