Remove deprecated `ITopicRepository.Save(isDraft)`
In OnTopic 2.x, we had introduced the idea of saving a draft, which was a specific version of a topic which wouldn't get published to the live site unless it was running in a special mode which loaded the draft. This was based on the idea of having a separate server for editing and previewing content, and was enabled using the legacy Web Forms `TopicsSection` configuration element.

This support was removed as part of OnTopic 3.x, but the `isDraft` overload remained as a vestigal interface element. To avoid confusiong, we're removing it. This is a breaking change, though as an optional parameter, we only expect that `ITopicRepository` implementations were calling it as a passthrough in order to maintain the interface contract.

This includes removing the parameter from XML Doc references as well.