Marked deprecated `GetContentTypeDescriptors()` overload as obsolete
The `GetContentTypeDescriptors(contentTypeDescriptor)` overload (7a5a243) was introduced to allow the content type descriptors cache to be updated at the same time as retrieving a content type by passing along a reference to a `ContentTypeDescriptor` and allowing it to override the local cache. While this built off of the familiar `GetContentTypeDescriptors()` method, it was a poorly thought through design, as the semantics don't make the behavior clear. Given that, it was deprecated and replaced with the `SetContentTypeDescriptors()` method to be more explicit about the functionality (78a9d3f).

This simply moves the overload from deprecated to obsolete such that it now throws an exception if used. In a future version, we will remove this entirely.