Marked deprecated `Topic.Description` property as obsolete
This was a convenience method intended to aid in retrieving a common—but not universal—attribute. As this is generally only needed in views, as part of page rendering, it is better served by the view model support accessed via the `ITopicMappingService`. If access to the description attribute is still needed via `Topic`, it can be accessed using `topic.Attributes.GetValue()` or `topic.Attributes.SetValue()` instead.

This simply moves the property from deprecated to obsolete so that it now throws an exception if used. In a future version, we will remove this entirely.