Marked `MapTopicRoute(IRouteBuilder)` as obsolete
In ASP.NET Core, there are two ways to configure a route: with `IRouteBuilder` or `IEndpointRouteBuilder`. Previously, `OnTopic.AspNetCore.Mvc` supported both, even though `IEndpointRouteBuilder` is preferred for ASP.NET Core 3+. We're now marking the `IRouteBuilder` method as obsolete.

This effectively ends support for OnTopic on older versions of ASP.NET Core which don't support `IEndpointRouteBuilder`—which is fine, since the release version of `OnTopic.AspNetCore.Mvc` has only ever supported ASP.NET Core 3+. (A beta version supported ASP.NET Core 2+, but was never released.)

This simply moves the property from deprecated to obsolete so that it now throws an exception if used. In a future version, we will remove this entirely.