Set default value of `ITopicRepository.Delete(isRecursive)` to `false`
Originally, `ITopicRepository.Delete()`'s `isRecursive` parameter was set to a default value of `false`, as it only makes sense to do a recursive delete if the user opts into it. Unfortunately, however, there was a long-standing bug where `Delete()` failed to validate whether or not there were children, and thus would _always_ do a recursive delete, regardless of whether or not `isRecursive` was set. Whoops.

This was fixed in OnTopic 4.5.0 (a0cc21c). In order to avoid breaking backward compatibility, however, the default was changed to `true` so that it continued to operate as it had before (1773113); otherwise this would have been a breaking change.

Having the default be more aggressive doesn't make much sense; users should opt in to the more invasive mode. As such, now that we're preparing for a major release, I'm reverting this back to its original default of `false`. Now, if a user attempts to `Delete(Topic)` without explicitly defining `isRecursive`, and that topic has children, an exception will be thrown, and the topic will not be deleted.