Migrated view model `List<>` properties to use `Collection<>`
Microsoft's design guidelines recommend that public interfaces use `Collection<T>` instead of `List<T>` (CA1002). The `List<T>` type is faster, and optimized for performance, but includes a lot of low-level members that most users don't need, while `Collection<T>` has more high-level convenience members that are more intuitive for most users.

As this was a breaking change, however, we didn't implement this as part of our implementation of Microsoft's updated Code Analysis (27e6940). Now that we're preparing for a major release, we're able to make these changes which could potentially break some implementations.

Notably, this will break implementations that specifically access the member as a `List<T>` (e.g., as a cast or assignment), or which call some of the `List<T>` specific methods, such as `AddRange()`. These calls were previously removed from elsewhere in the OnTopic library in preparation for this migration, but may still persist in some client implementations.