Updated `ITopicRepository.Load()` to use `uniqueKey` instead of `topicKey`
In a previous commit, we created a new function, `GetTopicIDByUniqueKey`, which allows a `TopicID` to be looked up based on its `UniqueKey` instead of its `TopicKey` (ea97f91). This effectively replaces the legacy `GetTopicID`, which would lookup the _first instance_ of a topic which had a given `Key`. Since multiple topics can have the same `Key`, however, this was error prone.

As a result, `GetTopicIDByUniqueKey` requires that a fully qualified `UniqueKey` (e.g., from `Topic.GetUniqueKey()`) be passed along, instead of simply a `Topic.Key`. As a result, this was not implemented on `ITopicRepository.Load(topicKey)` as that would be a breaking change; any call passing a single topic key would fail, unless that topic key happened to be in the root.

For this major release, we're finally updating this to use `GetTopicIDByUniqueKey`. To avoid confusion, the parameter name is also being updated from `topicKey` to `uniqueKey` to help disambiguate _which_ key is appropriate here.