Moved `RelatedTopicBindingModel` to the `OnTopic.ViewModels` library
Previously, `RelatedTopicBindingModel` was the only concrete model implemented in the core OnTopic library. Normally, we only need (or want) model abstractions in the main library, such as the underlying `IRelatedTopicBindingModel`. Since the core OnTopic library doesn't rely on `RelatedTopicBindingModel`, it's being moved to `OnTopic.ViewModels` under a new `BindingModels` namespace.

Note: The nomenclature of `OnTopic.ViewModels.BindingModels` is a bit goofy. If we introduce more standardized binding models we might want to revisit a more invasive change of migrating to something like `OnTopic.Models.ViewModels` and `OnTopic.Models.BindingModels`. For now, however, this is a minor annoyance for an outlier.