Moved `isIncoming` overloads on `RelatedTopicCollection` to `internal`
The `isIncoming` parameter is used to maintain referential integrity within the topic graph, and is not intended to be called by external implementors. Given this, the methods which use `isIncoming` have been marked as `internal` to avoid confusion.

As part of this, the optional `isDirty` parameter for `SetTopic()` has been moved to the public overload, since that _is_ required externally—and is actively used by `ITopicRepository` implementations.

Note that the (now) internal `SetTopic()` overload's parameter order has changed, to maintain consistency with the public version. As there's only one internal call to it, however, this is not expected to matter.