Moved `BindingModelValidator` to `OnTopic.Mapping.Reverse` namespace
The `BindingModelValidator` is only used, as the name suggests, for validating binding models, which in turn are only assessed as part of the `IReverseTopicMappingService`. Given that, it has been moved from the main `OnTopic.Mapping` namespace to the `OnTopic.Mapping.Reverse` namespace.

As this is an internal class, this has no external impact. (And, in fact, since it only has one caller which is already in the `OnTopic.MappingReverse` namespace, it also has no internal impact.)