Operate off of `ITopicBindingModel` type, not `*TopicBindingModel` name
Initially, the `DynamicTopicBindingModelLookupService` looked for all types whose names ended in `TopicBindingModel`. This is similar to how the `DynamicTopicViewModelLookupService` works.

Binding models are different, however. The `ReverseTopicService` expects binding models to implement the `ITopicBindingModel` interface. As such, we needn't rely on their name to identify their purpose. Further, implementors may wish to use the same classes for view models and binding models; by forcing a naming convention that is at odds with that used with view models, we prevent that from happening.

Given that, the `DyanmicTopicBindingModelLookupService` now looks for classes which implement `ITopicBindingModel`, regardless of whether or not they end with `TopicBindingModel`. In practice, today, we expect these to be identical. But this offers more flexibility to implementors.