Ensured `ArgumentNullException` included `nameof(parameter)`
Not all cases where an `ArgumentNullException` can be thrown include the parameter name. In some cases, the parameter name was included, but as a hard-coded string, or part of a more descriptive error message.

In these cases, the general preference is to use `nameof(parameter)` so that the name is validated against the actual parameter name, and mismatches will be detected if the parameter name changes.