Converted `ArgumentException` to more appropriate `InvalidOperationException`
Updated `RelatedTopicCollection.SetTopic()` to throw an `InvalidOperationException` instead of an `ArgumentNullException` for scenarios where `isIncoming` is set on a `RelatedTopicCollection` that isn't set to `_isIncoming`. This should never occur, as this is an `internal` overload and managed by the OnTopic library; if it does, however, it suggests a design flaw in the code.