Prefer `ArgumentOutOfRangeException` for range checks
When a method parameter is being validated against a range—e.g., to determine if it is a positive number, or falls within a particular range of numbers—prefer `ArgumentOutOfRangeException` over the more generic `ArgumentException`. As `ArgumentOutOfRangeException` derives from `ArgumentException`, this should not be a breaking cahnge, but will provide implementors to capture a more specific exception, should they choose.