Prefer `InvalidOperationException` in `SetRelationships()` for consistency
While the previous `ArgumentOutOfRangeException` is defensible, all other validation errors from `ReverseTopicMappingService` throw `InvalidOperationException`, and there's no reason to think an implementor would want to capture this exception individually, compared to any other exceptions. As such, for consistency, and to simplify handling exceptions, this is now standardized to an `InvalidOperationException`.

There is an argument to be made that we might want to establish a `TopicMappingException` class, potentially with more specific derived classes, to provide more granular control over error handling. We may reevaluate that in the future. For now, however, all mapping exceptions should share a common exception.