Prefer `ReferentialIntegrityException` when validating `ContentType`
If `ITopicRepository.Save()` is called with a `topic` whose `ContentType` cannot be found, an `ArgumentException` is thrown.

That's fine. But, in all other cases, when there are predictable exceptions occur within an `ITopicRepository` implementation then we throw a `TopicRepositoryException` or derivative so that error handling can catch all such cases. Further, we already have a `ReferentialIntegrityException` for cases where a topic reference can't be honored. A `ContentType` is, effectively, a reference to a `ContentTypeDescriptor`, so this seems appropriate to use here.