Corrected overloads for `RelatedTopicCollection`
In a previous set of commits, I moved the `RelatedTopicCollection` overloads which accepted `isIncoming` to be marked as `internal` (5e40470, 5706897). In setting up their public overloads, however, I inadvertantly introduced an infinite loop by a) failing to ensure they passed a value for `isIncoming`, and b) creating an ambiguous reference for the internal overload by making the `isIncoming` parameter optional. Whoops!

This fixes that issue by ensuring the `public` version of `RemoveTopic()` defaults to setting `isIncoming` to `false`, and the `internal` version of `RemoveTopic()`  doesn't assume a default for `isIncoming`.