Remove support for .NET Standard 2.0
.NET Standard 2.0 is needed to maintain support for .NET Framework (4.6.1+) as .NET Standard 2.1 only supports .NET Core 3+ and .NET 5+. Previously, we continued to target .NET Standard 2.0 in order to support the **OnTopic-MVC**, **OnTopic-WebForms**, and **OnTopic-Editor-WebForms** projects.

With the release of .NET 5, and the fact that we never released an **OnTopic-Editor-MVC** project, in preference for ASP.NET Core 3+, there is no compelling reason to continue to support the .NET Framework. If any implementations continue to target .NET Framework, they can continue to use OnTopic 4.x for now. That said, we definitely encourage implementors to upgrade to .NET 5+, now that it is released.