Removed legacy handling of `EditorType`
Previously, we maintained a compiler condition to handle `EditorType` differently between .NET Standard 2.0 and .NET Standard 2.1, due to the fact that .NET Standard 2.0 didn't support using a `StringComparison` with `EditorType`. Now that we're not supporting .NET Standard 2.0, we can get rid of this.

While I was at it, I migrated to the preferred `StringComparison.OrdinalIgnoreCase`.