Merged `AttributeDescriptor` and `AttributeTypeDescriptor`
OnTopic Editor 4.0.0 formalized the idea of attribute types as content types, thus allowing each attribute type to have its own data model in the Editor. Prior to this, attribute type specific configuration values had to be configured via a single text attribute which was parsed and applied by each `{AttributeType}ViewModel` (or, prior to that, `{AttributeType}.ascx`).

To continue supporting **OnTopic-Editor-WebForms**. the `AttributeDescriptor`'s `EditorType` and `ModelType` necessited complex logic, which was moved to a `ConfigurableAttributeDescriptor` in the **OnTopic-WebForms** project. As a result, the `AttributeDescriptor` acted as an abstract base class for both `ConfigurableAttributeDescriptor` as well as a simplified `AttributeTypeDescriptor` for the **OnTopic-Editor-AspNetCore** classes.

Since we're no longer supporting .NET Framework (422e6ed), there's no need to maintain this abstraction. Instead, we can merge the simplified `AttributeTypeDescriptor` logic into `AttributeDescriptor` and use that as the base class for all of the `OnTopic.Metadata.AttributeTypes` base clases (which represent the strongly typed attribute descriptors, and map to their corresponding content types).