Decoupled `FakeViewModelLookupService` and `TopicViewModelLookupService`
Previously, the `FakeViewModelLookupService` derived from `TopicViewModelLookupService`, such that it aggregated the types from both. That's a fine approach and continues to be supported. At this point, however, we additionally have the new `CompositeTypeLookupService` which allows multiple different `ITypeLookupService`s to be joined, even if they don't derive from one another.

While using the `CompositeTypeLookupService` isn't strictly necessary here, using it makes it easier to test it, while also better representing real world scenarios. For instance, in the next release of OnTopic, we'll be separating the `EditorViewModelLookupService` so that it doesn't inheriting from the `TopicViewModelLookupService`, thus allowing implementers the flexibility of whether or not they want to use the `OnTopic.ViewModels`.