Deleted unused `DateModified` column
The `DateModified` column provides a timestamp of when a record was created. This predated the `Version` column. Originally, the `Version` column was an `INT` and served a different purpose than `DateModified`. Later, however, `Version` was migrated to a `DATETIME` and effectively serves the same purpose as `DateModified`. Given this, `DateModified` is being removed. It's worth noting that no SQL or C# code relied on this, so this isn't expected to be a breaking change, unless implementers were using it for some reason.