Set `Version` default to use UTC date
Typically, the `Version` is set by the `CreateTopic` or `UpdateTopic` stored procedure. In a previous update, the `SqlTopicRepository.Save()` method was updated to ensure that the `Version` was set using `DateTime.UtcNow` so that the timestamp wouldn't vary based on the server's location or time settings. Attributes that rely on the default—such as those created manually in SQL without explicitly setting the `Version`—aren't able to take advantage of that. To ensure that the logic is synchronized between the client library and the SQL defaults, the default constraint for `Version` is now set to `GETUTCDATE()`.