Update `GetTopics` to use new(er) `GetTopicIDByUniqueKey` function
In a previous update, `SqlTopicRepository` was updated to use the new(er) `GetTopicIDByUniqueKey` function over the legacy `GetTopicID` function, thus operating off the `UniqueKey` instead of just finding the first instance of a key (4c3d30f). In that commit, we also ensured that `ITopicRepository.Load(string)` calls relied on a fully qualified path (e.g., `Root:Configuration:ContentTypes`).

In this update, we extend that to by the default when calling the `GetTopics` stored procedure directly, as well.

As with the previous commit, the associated parameter is renamed from `@TopicKey` to `@UniqueKey` to ensure this expected value is clear.