Delete legacy `GetTopicID` function
In a previous update, `SqlTopicRepository` was updated to use the new(er) `GetTopicIDByUniqueKey` function over the legacy `GetTopicID` function, thus operating off the `UniqueKey` instead of just finding the first instance of a key (4c3d30f). More recently, the same was applied to the `GetTopics` stored procedure (5706897).

With this, the legacy `GetTopicID` function is no longer needed; the `GetTopicIDByUniqueKey` function should always be used instead. This addresses flakey behavior in `GetTopicID` since it didn't allow callers to specify a unique topic, and could thus lead to bugs.