Rename `GetTopicIDByUniqueKey` to `GetTopicID`
Now that we've deleted the legacy (and flakey!) `GetTopicID` function in preference for the new(er) `GetTopicIDByUniqueKey` function, we can rename the `GetTopicIDByUniqueKey` to the unqualified name `GetTopicID`. This is a more intuitive name now that there's no longer a `GetTopicID` function (ff58a71).

As part of this, updated the reference in `SqlTopicRepository`, the `GetTopics` stored procedure, as well as the `OnTopic.Data.Sql.Database` documentation (`README.md`).