Include key attributes in `AttributeIndex`
When we return attributes from `GetTopics`, we don't include the key attributes—i.e., `Key`, `ContentType`, and `ParentID`, which are required to construct each `Topic`. That's because those are separately extracted from the `Attributes` table from the `TopicIndex` view, and thus not needed.

In practice, this often causes confusion, as it looks like they're missing when querying `AttributeIndex`. Given this, I'm updating the logic to return the key attributes, but then explicitly exclude them from `GetTopics`. That makes the logic a bit more intuitive and clear, instead of making `AttributeIndex` overoptimized for the business logic of `GetTopics`.

In the future, we may move key attributes to the `Topics` table at which point this will be irrelevant, as the separation will be enforced based on the data storage. Until then, however, this improves upon the current behavior.