Add `@TopicKey` and `@ContentType` to `CreateTopic`, `UpdateTopic`
In a previous commit, the key attributes of `Key`, `ContentType`, and `ParentID` were migrated from `Attributes` to `Topics`. This updates the `CreateTopic` and `UpdateTopic` stored procedure—as well as its call from `SqlTopicRepository`—to include `@TopicKey` and `@ContentType`. (`@ParentID` was already required for `CreateTopic`, and continues to be handled via `MoveTopic` for updates.)

Technically, the `UpdateTopic` stored procedure only requires the `@TopicKey` or `@ContentType` parameters if they have changed; it's smart enough to fall back to the existing value if they aren't passed. That said, it doesn't hurt anything to hard code them in absence of library support for tracking their state change.

These three attributes continue to be saved in the `Attributes` table for now, and are thus maintained in two locations. That will be removed in subsequent commits.