Updated SQL code to get key attributes from `Topics`
The `Key`, `ContentType`, and `ParentID` are currently being maintained in both `Attributes` (as they've historically been) and `Topics` (c50f44d, 1124443, a74389d). Pulling data from `Topics` is dramatically easier since a) the data is in columns, b) the columns are strongly typed, and c) there is no versioning to deduplicate. This is important as the key attributes are frequently relied upon by other functions and procedures as part of joins. Given this, we can now simplify those queries to instead rely upon the `Topics` table instead of doing a complex query against `Attributes` or a join against `TopicIndex`.

For now, the `Key`, `ContentType`, and `ParentID` attributes continue to be stored in `Attributes`; in a future commit, we'll be removing that as well.