Introduce `Topic.IsDirty()` for state tracking
In previous commits, we introduced `AttributeValueCollection.IsDirty()` (1560a46) and `RelatedTopicCollection.IsDirty()` (2adef58) for tracking changes. This allowed us to intelligently determine whether or not a `Topic` needed to be saved as part of a recursive `ITopicRepository.Save()`, thus speeding up bulk imports and updates (afa1e9b).

As we've now moved `Key`, `ContentType`, and `Parent` out of `Attributes` and to `Topic`, we need an additional level of state tracking, as otherwise we won't be able to detect if e.g. the `Key` and/or `ContentType` have changed unless other attributes or relationships have _also_ changed. To do this, we've introduced a `Topic.IsDirty()`.

Given its centralized position, however, we've also added an overload to `Topic.IsDirty()` to `checkCollections`—which will call `Topic.Attributes.IsDirty()` and `Topic.Relationships.IsDirty()`—and to optionally `excludeLastModified`, which relays that parameter to `Topic.Attributes.IsDirty()`.

This logic is integrated with `SqlTopicRepository` to ensure that it will save a topic even if the only change is the `Key` or `ContentType`. (There was already handling for `Parent` under `ITopicRepository.Move()`.)

Finally, three unit tests have been introduced for verifying three different states of `Topic.IsDirty()`—i.e., that a new topic, or a topic whose `Key` or `ContentType` properties are changed is flagged as `IsDirty()`, while an existing topic (with an `Id`) is not.