Delete core attributes from `Attributes` as part of pre-deployment script
In a previous commit, a basic pre-deployment migration script was established which created the `TopicKey`, `ContentType`, and `ParentID` columns in the `Topic` table and copied the core attributes from `Attributes` to `Topics` (1124443).

Now that the migration is mostly complete, and all functions and stored procedures (2dc4fb2) and business objects (ea25ed3) are correctly saving that data to `Topics`, we can safely delete the legacy data from `Attributes`.