Removed now-obsolete `TopicIndex` view
The `TopicIndex` extracted the latest version of the key attributes—i.e., `Key`, `ContentType`, and `ParentID`—from the `Attributes` table and pivoted them to columns. That complexity can now be handled by a straight query of the `Topics` table, as has already been implemented in the dependency functions and stored procedures (2dc4fb2). As such, the `TopicIndex` view is no longer used, and no longer serves a purpose, and can be safely removed.