Updated `AttributeValueCollection` tests to use `View` for business logic tests
The `AttributeValueCollection.EnforceBusinessLogic()` method looks for corresponding properties on the associated `Topic` instance which may have a property with the same name as an `AttributeValue` being modified. If so, it routes the user through that property to ensure that any business logic and state management is maintained.

Previously, the `AttributeValueCollection`'s unit tests relied on the `Topic.Key` property as the quentissential example of this, since it was not only an `[AttributeSetter]`, but also enforced business logic, by ensuring that the value was a proper key.

With the `Key` now operating off of a local field (ea25ed3), those tests are no longer valid. As such, they've been updated to instead use the `Topic.View` field which enforces similar logic, and remains used today.