Simplified `isDirty` tracking in `SqlTopicRepository.Save()`
Previously, the call to `Topic.Attributes.IsDirty()` was set to conditionally set `excludeLastModified` based on the state of `areRelationshipsDirty`.

There's some logic to that. But because both `areAttributesDirty` and `areRelationshipsDirty` (and, now, `isTopicDirty`) are all evaluated as part of the determination of whether or not to save the topic, there's no real need to add the complexity of a conditional `excludeLastModified`—we can always set that to `true`.

This is much easier to reason through that the weird double-negative created by `excludeLastModified: !areRelationshipsDirty`, and ends up with the same end result. I.e., we also only care if `Attributes` are dirty if `LastModified` and `LastModifiedBy` aren't the _only_ attributes that were modified.