Update `[FilterByAttribute()]` tests to not use `ContentType`
Since the `ContentType` is no longer stored as an `Attribute`, the `[FilterByAttribute()]` unit test fails, as the `FilteredTopicViewModel` was filtering by `ContentType`. To mitigate that, the unit test is updated to filter by two arbitrary attributes instead.

In the future, we'll need to reestablish a way to filter by `ContentType`, as that's actually the most common use case for the `[FilterByAttribute()]` attribute. But, for now, this ensures all of our unit tests are passing, and fully accounting for the change of location for `ContentType`.