Introduced new `[FilterByContentType()]` attribute
Since `ContentType` has been moved from `Topic.Attributes` to `Topic.ContentType`, it can no longer be filtered using the `[FilterByAttribute()]` attribute during mapping (4a9f5b7). That's a big deal since that's the primary use case for the `[FilterByAttribute()]` attribute!

To mitigate that, a new `[FilterByContentType()]` attribute is introduced, read by the `PropertyConfiguration`, and enforced by the `TopicMappingService`. In addition, a unit test was introduced to validate the functionality.

(As part of this, an old unit test had to be renamed as this inadvertantly introduced an ambiguity with the test evaluating a collection filtered based on the collection type.)