Establish basic pre-deployment migration script for 4.0.0
This is a bit late, but as we've established a pre-deployment script for migrating from 4.0.0 to 5.0.0, I'm also committing our legacy pre-deployment script for migrating from 3.0.0 to 4.0.0. This was maintained independently by Ignia, but really makes more sense as part of the `OnTopic.Data.Sql.Database` project.

This one is most notably needed to migrate the legacy generic `AttributeDescriptor` references to their specific `AttributeDescriptor` descendents. E.g., a text attribute no longer uses the generic `AttributeDescriptor` content type, but instead uses the strongly typed `TextAttribute` content type.

As part of this, it also ensures the `ContentType` is explicitly set on each `Topic`, whereas previously these values could be inherited from a `DerivedTopic`. While technically fine, that introduces confusion with reporting and updates like this one, so it's easier all around to ensure those are repeated. (This business preference is firmly enforced in OnTopic 5.0.0 due to the `ContentType` being moved to the `Topics` table.)