Skip including of `PageGroup` as part of the `SitemapController`
When writing the sitemap, the `SitemapController` skips over any `Container` content types—but still includes their descendents. We want to add `PageGroup` to that list, as well as any topics with a `Url` attribute, since those will always redirect the request to another page, and thus shouldn't be in the sitemap.

To facilitate this, I refactored the `Container` and `PageGroup` into a new `SkippedContentTypes` list, similar to the existing `ExcludeContentTypes` and `ExcludeAttributes` lists. That will make it easier in case we want to add additional content types to skip in the future.

Finally, updated the unit tests to account for this new scenario.