Unhide unnecessary attributes from `SitemapController`
`IsActive` is not used, and should not ever be present—at least as a global attribute. `Url` is used, but a new feature skips any topics that have it defined (a90fe9f), and so we'd never expect it to be present.

Technically, `IsDisabled` and `NoIndex` shouldn't show up, as they're also excluded from the `SitemapController`, but they might show up as a `0` if they're explicitly disabled, and so they're being maintained. I'm not sure this is necessary, but it doesn't hurt anything, and helps reduce the amount of clutter.