Implemented `TypeLoadException` if target DTO class is invalid
If the `ITypeLookupClass` can't find a type associated with a particular content type, then it used to throw an `InvalidOperationException`, followed recently by a `TopicMappingException`.

The problem with these is that they're ambiguous, and could represent a number of _other_ issues. Since this exception represents a potentially serious design-time error, I'm replacing it with the more specific `TypeLoadException`.

As part of this, I'm also updating the `catch` statement in `SetTopicReferenceAsync()`, which is specifically looking for these errors with a previously imprecise catch statement.

In the future, we _may_ want to replace this error with a specific type—e.g., `MappingTypeException`—that is a subclass of the new `TopicMappingException`. That would be more consistent with the original concept of that class. We'll reevaluate that as part of our broader reassessment of exceptions. For now, however, this at least provides a specific exception that can be caught regarding this exact error.